context menus move to widgets
