+++ src/ducks/Watchlists/Actions/Edit/Trigger.module.scss
