+++ src/ducks/Watchlists/Widgets/TopPanel/WatchlistActionButton.module.scss
@@ -1,0 +1,11 @@
+@import "~@santiment-network/ui/variables.scss";
+
+.icon {
+  height: 16px;
+  width: 16px;
+  fill: var(--casper) !important;
+}
+
+.text {
+  margin-left: 7px;
+}
