+++ src/ducks/Watchlists/Widgets/TopPanel/WatchlistActions.js
@@ -3,15 +3,15 @@ import { CSVLink } from 'react-csv'
 import { graphql } from 'react-apollo'
 import Icon from '@santiment-network/ui/Icon'
 import Button from '@santiment-network/ui/Button'
-import { normalizeCSV } from '../utils'
-import { isNotSafari } from '../../../utils/utils'
-import { upperCaseFirstLetter } from '../../../utils/formatting'
-import ShareModalTrigger from '../../../components/Share/ShareModalTrigger'
-import EditTrigger from './Edit/Trigger'
-import WeeklyReportTrigger from './WeeklyReport/Trigger'
-import Copy from './Copy'
+import { normalizeCSV } from '../../utils'
+import { isNotSafari } from '../../../../utils/utils'
+import { upperCaseFirstLetter } from '../../../../utils/formatting'
+import ShareModalTrigger from '../../../../components/Share/ShareModalTrigger'
+import EditTrigger from '../../Actions/Edit/Trigger'
+import WeeklyReportTrigger from '../../Actions/WeeklyReport/Trigger'
+import Copy from '../../Actions/Copy'
 import WatchlistContextMenu from './WatchlistContextMenu'
-import { WATCHLIST_QUERY } from '../../../queries/WatchlistGQL'
+import { WATCHLIST_QUERY } from '../../../../queries/WatchlistGQL'
 import styles from './WatchlistActionButton.module.scss'
 
 const WatchlistActions = ({
