+++ src/ducks/Watchlists/Widgets/TopPanel/WatchlistContextMenu.js
@@ -4,12 +4,12 @@ import ContextMenu from '@santiment-network/ui/ContextMenu'
 import Icon from '@santiment-network/ui/Icon'
 import Button from '@santiment-network/ui/Button'
 import Panel from '@santiment-network/ui/Panel/Panel'
-import { normalizeCSV } from '../utils'
-import Delete from './Delete'
-import Edit from './Edit'
-import WeeklyReport from './WeeklyReport'
-import Copy from './Copy'
-import VisibilityToggle from './ChangeVisibility'
+import { normalizeCSV } from '../../utils'
+import Delete from '../../Actions/Delete'
+import Edit from '../../Actions/Edit'
+import WeeklyReport from '../../Actions/WeeklyReport'
+import Copy from '../../Actions/Copy'
+import VisibilityToggle from '../../Actions/ChangeVisibility'
 import styles from './WatchlistContextMenu.module.scss'
 
 const WatchlistContextMenu = ({
