+++ src/ducks/Watchlists/Widgets/TopPanel/WatchlistContextMenu.module.scss
