+++ src/pages/Watchlists/AssetsMobilePage.js
@@ -12,7 +12,7 @@ import { getWatchlistName } from '../../ducks/Watchlists/utils'
 import { addRecentWatchlists, removeRecentWatchlists } from '../../utils/recent'
 import AssetCard from '../../ducks/Watchlists/Widgets/Table/AssetCard'
 import AssetsTemplates from '../../ducks/Watchlists/Widgets/Table/AssetsTemplates'
-import WatchlistActions from '../../ducks/Watchlists/Actions/WatchlistActions'
+import WatchlistActions from '../../ducks/Watchlists/Widgets/TopPanel/WatchlistActions'
 import WatchlistAnomalies from '../../ducks/Watchlists/Widgets/WatchlistOverview/WatchlistAnomalies/WatchlistAnomalies'
 import PageLoader from '../../components/Loader/PageLoader'
 import MobileHeader from './../../components/MobileHeader/MobileHeader'
